# Fuse Elemwise graphs that have multiple outputs and clients