… tests/tensor/rewriting/test_math.py
… tests/tensor/rewriting/test_subtensor.py
… tests/tensor/test_subtensor.py
… tests/test_printing.py