Continuation of https://github.com/aesara-devs/aesara/pull/1242, as I've been blocked on that repo

More context can be found in https://github.com/aesara-devs/aesara/issues/1237

Clarifying what is the idea behind allowing Composites with multiple clients. The following example:

```python
import pytensor as aesara
import pytensor.tensor as as at

x = at.vector("x")
y = at.exp(x/5)
w = y + 1
z = y * 2

f1 = aesara.function([x], [z, w])
aesara.dprint(f1)
```

Before this PR it produced this graph:
```
Elemwise{Mul}[(0, 1)] [id A] 2
 |TensorConstant{(1,) of 2.0} [id B]
 |Elemwise{Composite{exp((i0 * i1))}} [id C] 0
   |TensorConstant{(1,) of 0.2} [id D]
   |x [id E]
Elemwise{add,no_inplace} [id F] 1
 |TensorConstant{(1,) of 1.0} [id G]
 |Elemwise{Composite{exp((i0 * i1))}} [id C] 0
```

After:
```
Elemwise{Composite{(i2 * exp((i0 * i1))), (i3 + exp((i0 * i1)))}}.0 [id A] 0
 |TensorConstant{(1,) of 0.2} [id B]
 |x [id C]
 |TensorConstant{(1,) of 2.0} [id D]
 |TensorConstant{(1,) of 1.0} [id E]
Elemwise{Composite{(i2 * exp((i0 * i1))), (i3 + exp((i0 * i1)))}}.1 [id A] 0
```

The expected savings come from having to iterate only once over the data vector x, vs having to iterate 3 times before, once over x and twice over `exp(i0 * i1)`.

Another example is the following graph:
```python
x = at.dvector("x")
mu = at.dvector("mu")
logp = (- ((x - mu) **2) / 2)
grad = at.grad(logp.sum(), x)
f2 = aesara.function([mu, x], [logp, grad])
aesara.dprint(f2)
```

Before:
```
Elemwise{Composite{(i0 * sqr(i1))}}[(0, 1)] [id A] 2
 |TensorConstant{(1,) of -0.5} [id B]
 |Elemwise{sub,no_inplace} [id C] 0
   |x [id D]
   |mu [id E]
Elemwise{neg,no_inplace} [id F] 1
 |Elemwise{sub,no_inplace} [id C] 0
```

After:

```
Elemwise{Composite{(-(i0 - i1)), (i2 * sqr((i0 - i1)))}}.1 [id A] 0
 |x [id B]
 |mu [id C]
 |TensorConstant{(1,) of -0.5} [id D]
Elemwise{Composite{(-(i0 - i1)), (i2 * sqr((i0 - i1)))}}.0 [id A] 0
```

Again we replace 3 loops by 1.

For more detail, here is the C code of the Composite scalar in the first function:

```C
// function 1
{
npy_float64 V13_tmp1;
// i0 * i1
V13_tmp1 = V11_i * V9_i;

npy_float64 V13_tmp2;
// exp(i0 * i1)
V13_tmp2 = exp((npy_float64)V13_tmp1);

// First output
// i2 * exp(i0 * i1)
V1_i = V7_i * V13_tmp2;

// Second output
// i3 + exp(i0 * i1)
V3_i = V5_i + V13_tmp2;
}
```

```C
// function 2
{
npy_float64 V11_tmp1;
// i0 - i1
V11_tmp1 = V9_i - V7_i;

// First output
// - (i0 - i1)
V1_i = -V11_tmp1;

npy_float64 V11_tmp2;
// sqr(i0 - i1)
V11_tmp2 = V11_tmp1 * V11_tmp1;

// Second output
// i2 * sqr(i0 - i1)
V3_i = V5_i * V11_tmp2;
}
```

You can see that it avoids recomputing the same sub-expressions in both cases. It does store more intermediate variables than needed, but hopefully the compiler will be able to get rid of those.

This would be further improved by inplacing the scalars as is being pursued in #107 

## TODO:
- [x] Wait for Numba to allow multi-output Elemwise, or figure out how to disable rewrite in that backend
- [x] Figure out what to do with inplace optimization when there are multiple outputs
  -  Deferred as a separate issue: https://github.com/pymc-devs/pytensor/issues/138 
- [x] Refactor the specialized `add_mul_fusion`
- [x] Do not split fusion ops depending on c-code presence in non-c backends
  - [x] Similarly, get rid of fragile `elemwise_max_input_fct`
- [x] Improve Op string, it now looks like a monster with many outputs
- [x] Profile expected performance gains!  # Will do after #139 

## New features
- Fuse Elemwise graphs that have multiple outputs and non-fuseable clients

## Bugfixes
- Fix bug when creating Composite with multiple identical outputs