@ricardoV94 2022-12-13
@OriolAbril Any idea why readthedocs is failing?