@OriolAbril 2022-12-13
I suspect github is laggy or partially down. The error was that readthedocs was unable to find the reference to this PR. I restarted the job and now it has been able to get the PR changes. 🤷🏿