@ricardoV94 2022-12-14
Yup, the slow and skip marks seem to be ignored in the CI [as hinted in #](https://github.com/pymc-devs/pytensor/issues/5#issuecomment-1350447203)