@ricardoV94 2022-12-14
Alright first time the tests pass (minus the ones I disabled :D)