@ricardoV94 2022-12-15
Benchmark, running this snippet after this PR vs main:

```python
import pytensor
import pytensor.tensor as pt
import numpy as np

rng = np.random.default_rng(123)
size = 100_000
x = pytensor.shared(rng.normal(size=size), name="x")
mu = pytensor.shared(rng.normal(size=size), name="mu")

logp = (- ((x - mu) **2) / 2)
grad = pt.grad(logp.sum(), x)

func = pytensor.function([], [logp, grad])
pytensor.dprint(func)
%timeit -n 1000 func()
```

This PR: 145 µs ± 15.6 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
Main: 228 µs ± 26.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

Weirdly, it's always ~2x faster the second time I run the script (in the same IPython console). No idea what's going on with that