@aseyboldt 2022-12-15
Timings on my desktop machine give a similar pattern.
![image](https://user-images.githubusercontent.com/1882397/207986163-b592237f-2b39-4b28-b05f-431916b1c590.png)