@ricardoV94 2023-01-18
This PR is not merging subgraphs that are only connected via a non-mergeable input such as 

```python
x = pt.vector("x")
y1 = pt.exp(pt.sin(x))
y2 = pt.sin(pt.exp(x))
pytensor.function([x], [y1, y2])
```

That is added as an xfail test for now. I consider that an improvement we can do in a follow-up