@ricardoV94 2023-01-18
Added example above as a benchmark test for C and Numba backends. Both shows a clear improvement (Numba more so than C)