@ricardoV94 2023-01-18
Now that Numba accepts multi-output Elemwise, I think this is ready for review