@ricardoV94 2023-02-07
The refactor commit shows up as WIP just to see if the tests pass. I will merge it with the commit where the changes are made if it does