@ricardoV94 2023-02-08
> thanks for cleaning up the code some more, @ricardoV94
> 
> I know that you invested a lot of attention on the details here. If there is anything where you're in doubt and need a careful review let me know. It's just too much for one pass..

Yeah, it's a though one to review. Thanks for the suggestions to make it a tiny bit more readable. I think I will merge it and iterate in future PRs